refactor position
