+++ apps/web/preload/script/api/index.ts
@@ -1,4 +1,4 @@
-import { processDom } from './dom';
+import { captureScreenshot, processDom } from './dom';
 import {
     getChildrenCount,
     getElementAtLoc,
@@ -22,7 +22,6 @@ import { getElementIndex, moveElement } from './elements/move';
 import { drag, dragAbsolute, endAllDrag, endDrag, endDragAbsolute, startDrag } from './elements/move/drag';
 import { getComputedStyleByDomId } from './elements/style';
 import { editText, isChildTextEditable, startEditingText, stopEditingText } from './elements/text';
-import { captureScreenshot } from './events/dom';
 import { handleBodyReady } from './ready';
 import { setFrameId } from './state';
 import { updateStyle } from './style';
